rename old implementation to legacy and new to standard
